correct the Text property of the VeQItem
The Text property was defined as a QVariant with a valueChanged as notify
signal, likely because it got incorrectly copied from the value property.
Make it a QString instead with the textChanged signal for notification.

This shouldn't have any impact, in the declarative world the VeQuickItem
is used instead, which does subscribe to the textChanged signal. In the
c++ world the textChanged signal would directly be subscribed to as well.