ve_qitem: keep the last valid values
If a service disconnects, the behaviour of the values depended on the
application. Either all values with kept or everything was invalidated.
However in practice it depends on the kind of value, e.g. product name,
firmware versions etc of the missing device you might want to keep /
show to explain what went missing. On the otherhand outdated runtime
values should not be shown.

This replaces the treewide setting with making both of them available,
the current value and the last valid value. An application can then
decide per value if it wants to show the current value or the last
valid one.

It also makes the impementation of a backend more straight forward,
since it should simply always invalidate the current value when it
is no longer available.