No longer service MQTT message queue asynchronously
Qt For WebAssembly doesn't process the events in a timely manner
(unless the event queue is kicked by a timer event), so abandon
the asynchronous queue.  This may be fixed in a future version of
Qt, at which point in time it may be beneficial to revert this
commit, if stutters/UI stalls are noticed when receiving large
batches of MQTT notifications.