Invalidate MQTT item values on disconnect / null
The client now has the option whether to read the last-valid value
or the most-recent (even if invalid) value.

To support this, we need to ensure that we produce a
specified-as-invalid value whenever connection to the backend is
lost, or when the backend tells us that a specific device has been
disconnected (via a null payload message for a topic).