Fix websocket disconnection handling
Qt MQTT already emits the appropriate disconnected signal and
closes the websocket when it detects that the websocket has
disconnected, so the previous manual handler caused the signals
to be emitted twice.

However, in non-websocket case, we also need to handle the case
where the initial connection fails (and manually trigger our
onDisconnected() handler, to perform reconnect etc).
We detect this by handling error change and state change from
the MQTT client.

Finally, fix the reconnect count (as it was previously reset in
a queued connection).