Transition to Ready state after all initial messages are received
Previously, we transitioned to Ready state as soon as we received
the first message (since that meant that communication with the
broker had been established).

Now we transition to Initializing state once we receive the first
message, and only transition to Ready state once we have received
all of the initial messages triggered by the first KeepAlive.

This ensures that data is in a coherent state when Ready.

Also, construct timers as children of the producer, to ensure that
timer events are serviced on the producer's thread.