veqitem: make it possible to poll values
Normally values are updated by the sender by sending change events and cached
on the receiving side. If the values is cached, it is not requested. However,
for some information, CAN-bus network statistics e.g., it is of hardly any
interest on a properly running device and it is a rather waste of cpu cycles
to get them continiously and send them around all the time.

The dbus code can already request a value even though a cached version is
available. This makes the available in the base class and the VeQuickItem, so
UI code can poll such values on a periodic based only when the values are shown
to the user.