VeQItem: add using statements
To allow forced retrievals (as in never report a cached value) and overload
was introduced and the original function calls that with forced set to false.
According to Woverloaded-virtual those the base class members are not available
in the derived classes. Add the using statement to make them available again.

This doesn't solve an actual problem afaik, only the warning.