Close websocket device upon disconnection
Safari will disconnect the websocket when the tab is idle, in
order to save power etc.  When this occurs, we need the device
to recognise this case and close() the socket (client side),
so that the QMqttClient detects the new IO device state, and
transitions properly to Disconnected state.

Then, the UI code can detect the state transition, and rebuild
the data connection and the UI appropriately.