add bms & charger error classes
At the moment these are classes, so it can be used from QML as well. Support
for that can be dropped, but it doesn't harm either.

NOTE: The error codes are defined at the sending side, they should NOT be added
here at the receiving side when not being defined at the sending side!!