Don't crash if given malformed uid string from untrusted source
If the QML code sets some malformed UID string, its parts might
not match a valid UID path, and attempting to dereference children
matching the path parts will crash.

Instead of crashing, check to make sure that each path part
corresponds to a valid object, and return nullptr if not.