MQTT: ensure non-leaf items never go into VeQItem::Offline state
When the MQTT producer is disconnected, leaf items should have their
states set to VeQItem::Offline. This should not be done for non-leaf
items, because when the producer is reconnected, it will only receive
updated values for leaf items; therefore, if a non-leaf item had its
state set to VeQItem::Offline, it would stay offline and never return
to the Idle or Synchronized states.

This fixes a bug where if a VeQItemSortTableModel had non-leaf uids
and was set to FilterOffline, the non-leaf uids would be filtered out
when the MQTT transport was disconnected, but when reconnected, the
non-leaf uids would stay offline and would not be added back to the
model.

https://github.com/victronenergy/gui-v2/issues/1666