ve_quick_item: send the correct type in SetValue
The javascript engine isn't strictly typed and might convert an int to
a double, especially when an operation is performed on it. But it seems
even round returns a double in Qt6.

Services aren't supposed to be too picky about accepting values if they
can trivially convert it, but since this is since Qt6, there are bugs
in there. So this is the "be carefull what you send" part.

When an unit converter is used, this now also sends the same type
back as was reiceived instead of a double.

https://github.com/victronenergy/venus-private/issues/558