VeQItemMqtt: use nlohmann/json to correctly parse floats
The QtJson turns floats which happen to have an integer value into
integers. This causes problems since the code now makes sure the
same type is written back as received, which then starts to round
floats into integers when being changed.

Since nlohmann/json doesn't have this problems, use that to
deserialize any basic value (the common case). Fall back to qt for
now to handle dict and arrays, which is seldomly the case, but care
should be taken that the QVariant / qml understand that type. And
preferably without making additional copies of the arrays / dicts.