VeQItem: forward the setValue result to QML
Let QML know what the result of a setValue is / wether an error
occured. Note that this is only usefull for localsettings and
venus-platform etc. When data is forwarded to a product, most
services will only report that the request has been accepted,
but it might not yet have been done.